ENGR00126584 usb gadget driver: Clarify some usb handling routines
1. When the usb is not in use, the usb can be put into low power mode
(usb related clock are off, and phcd = 1)
2. When the usb is in low power mode, the usb wakeup irq can be triggered,
and let the usb return to normal mode
3. The flag device_can_wakeup(struct device *dev) is standing for the device
has(and will use) the abilities for waking up the whole system, not means
the device has the abilities for waking up itself.
4. During the pm process, wakeup is only needed to enable when the user
needs usb device to wakeup the whole system.
5. The initialization changes a little, for evk board is not connected to pc
as an example:
Loading the driver:
1. Clear phcd and Open clock
2. init usb registers
3. vbus is not valid, and go to low power mode
connect usb line to pc:
1. Wakeup irq happens
2. usb goes to normal mode
3. vbus irq happens(at the same interrupt)
4. set run_stop bit
5. enumeration begins

Signed-off-by: Peter Chen <peter.chen@freescale.com>
Signed-off-by: Huhui <b29976@freescale.com>