ENGR00137667-1 usb-otg: refine the otg configration
CONFIG_USB_OTG & CONFIG_USB_OTG_UTILS is better to be only
selected by CONFIG_MXC_OTG. In that case, the user only
needs to select/de-select one config to enable/disable usb otg configuration.

However, the user needs to disable host or device function for otg port
if the user does not choose CONFIG_MXC_OTG. The reason for that is
there are two usb drivers for host and udc, but they uses the same resources.
So, we only needs to use one driver for non-otg mode.

Reported-by: Robby Cai <r63905@freescale.com>
Signed-off-by: Peter Chen <peter.chen@freescale.com>