ENGR00139196 MX53 ARD: add VGA support
The VGA pins in MX53 ARD share the same PINs with WEIM.
So when the VGA is used, the ethernet and NOR via WEIM can
not be used. To run VGA on MX53 ARD, please do as the following:
- add boot option "ard-vga" in kernel launch command line.
- Flash the images on SD or NAND and avoid to use Ethernet.
- HW configuration: Short PIN 1, 2 for J11 and J12 in CPU board.
Please note: If using the ethernet, Short PIN 2, 3 for J11 and
J12.

Signed-off-by: Lily Zhang <r58066@freescale.com>