ENGR00139192 - EPDC fb: Refine merging and workaround missed collision
Added code to detect the potential for a "missed collision", where
an update would cause a collision with an update sitting on the collision
list, but does not do so with currently active updates. For this case,
we conservatively treat the update as if it has generated a collision.

Refined merging algorithm by blocking the merge process upon
detecting certain conditions where no further merges should be done.
Also added check for a special case where update regions are identical.

Signed-off-by: Danny Nold <dannynold@freescale.com>